+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/papify/PapifyPage.java
@@ -64,10 +64,8 @@ import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Text;
 import org.eclipse.swt.widgets.TreeColumn;
 import org.eclipse.ui.IEditorPart;
-import org.eclipse.ui.IPropertyListener;
 import org.eclipse.ui.forms.IManagedForm;
 import org.eclipse.ui.forms.editor.FormEditor;
-import org.eclipse.ui.forms.editor.FormPage;
 import org.eclipse.ui.forms.events.ExpansionAdapter;
 import org.eclipse.ui.forms.events.ExpansionEvent;
 import org.eclipse.ui.forms.widgets.ExpandableComposite;
@@ -91,13 +89,14 @@ import org.preesm.model.slam.utils.DesignTools;
 import org.preesm.ui.scenario.editor.FileSelectionAdapter;
 import org.preesm.ui.scenario.editor.Messages;
 import org.preesm.ui.scenario.editor.PreesmAlgorithmTreeLabelProvider;
+import org.preesm.ui.scenario.editor.ScenarioPage;
 
 /**
  * Papify editor within the implementation editor.
  *
  * @author dmadronal
  */
-public class PapifyPage extends FormPage implements IPropertyListener {
+public class PapifyPage extends ScenarioPage {
 
   /** Currently edited scenario. */
   private Scenario scenario = null;